improvement(continuous event): move ContinuousEvent to separate module
Move ContinuousEvent, ContinuousEventsRegistry and ContinuousRegistryFilter to the its own
module.
Also move BaseStressEvent that import from ContinuousEvent, and all connected classes to
the separate module.
Update modules that use ContinuousEvent accordingly.
