+++ sdcm/sct_events/nodetool.py
@@ -15,7 +15,7 @@ from json import JSONEncoder
 from typing import Any, Type
 
 from sdcm.sct_events import Severity
-from sdcm.sct_events.base import ContinuousEvent
+from sdcm.sct_events.continuous_event import ContinuousEvent
 
 
 @dataclass
